mv deep fool
